Add files via upload
Stocks_nn takes in the data and labels from stocks_preprocessing_for_nn and creates a dataset used to train it's neural network. The neural network trains slowly, and at 50 epochs it is a bit worrying  that I am overtraining it on the same data. However, the learning rate is set extremely low so that it can find the subtleties in stock technical data.

This neural network aims to predict the movement of the stock in the coming 14 day, and is tested using the most recent 40 days of technical data.